requirements.txt
pillow==11.0.0
pytube==15.0.0
tk==0.1.0
